Add testing and benchmarking function for sorting algorithms
The commit introduces a testing function for evaluating the performance of different sorting algorithms. This function, 'perform_tests', uses timeit to measure the execution time of each algorithm. The testing occurs on three kinds of datasets - ordered, reverse and random, generated by a new 'generate_data' function in the 'data_generator.py' file.